[~] BuildTable hooks
