+++ src/components/UI/Layout/HKubeLayout.css
@@ -2,3 +2,8 @@
   background-color: white;
   color: black;
 }
+
+.ant-table-fixed-header .ant-table-scroll .ant-table-header {
+  overflow: hidden !important;
+  margin-bottom: 0px !important;
+}
