+++ src/components/dumb/StatusTag.react.js
@@ -8,7 +8,7 @@ import { STATUS } from 'constants/colors';
 function StatusTag({ status, count }) {
   return (
     <Tooltip placement="top" title={toUpperCaseFirstLetter(status)}>
-      <Tag color={STATUS[status]}>{[count]}</Tag>
+      <Tag color={STATUS[status]}>{count}</Tag>
     </Tooltip>
   );
 }
